test(compiler): u32_arithmetic uses unsigned long after parent A1
Companion fix for the parent OpenSNES SDK chantier A1 (commit 6eb62e1
on develop). The parent shipped cproc IR int=2, long=4 — moving the
canonical 32-bit type from `unsigned int` (was 4 bytes pre-A1) to
`unsigned long` (now 4 bytes post-A1).

This fixture was the only test in the compiler-test phase that hard-
coded `typedef unsigned int u32` locally with the assumption that
`unsigned int = 4 bytes`. After A1 it becomes 2 bytes; the
`result32 is dsb 2, expected dsb 4` assertion fires correctly.

The fix mirrors the parent's `lib/include/snes/types.h` update —
`u32` typedefs to `unsigned long` rather than `unsigned int`. Comment
block updated to document the post-A1 type sizes.

Validation: full quick-suite (`--allow-known-bugs` per CI config)
returns 404/404 under the parent's A1-shipped develop. The
`u32_arithmetic` test now correctly verifies multi-word arithmetic
for the new 32-bit type.

+++ test/fixtures/compiler/test_u32_arithmetic.c
@@ -1,16 +1,17 @@
 // =============================================================================
-// Test: 32-bit (unsigned int) arithmetic
+// Test: 32-bit (unsigned long) arithmetic
 // =============================================================================
 // Prevents: Incorrect multi-word arithmetic on 16-bit CPU
 //
 // The 65816 is a 16-bit CPU. 32-bit operations require multi-word
 // sequences (two 16-bit operations with carry propagation).
 //
-// NOTE: On cproc/65816, type sizes are:
+// NOTE: On cproc/65816, type sizes (since chantier A1, 2026-05-08):
 //   unsigned char  = 1 byte (u8)
 //   unsigned short = 2 bytes (u16)
-//   unsigned int   = 4 bytes (u32)  <-- THIS is 32-bit
-//   unsigned long  = 8 bytes (u64)  <-- NOT 32-bit!
+//   unsigned int   = 2 bytes (native 16-bit word)
+//   unsigned long  = 4 bytes (u32)  <-- THIS is 32-bit
+//   unsigned long long = 8 bytes (u64)
 //
 // Uses function parameters to prevent constant folding, ensuring
 // the compiler generates actual runtime arithmetic instructions.
@@ -18,7 +19,7 @@
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+typedef unsigned long u32;
 
 u32 result32;
 u16 result16;
